ProPresenter/requirements.txt
requests
pyyaml
argparse
validators
